+++ NEWS.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Allow setting `slot` parameter in `RunUMAP`
 - Added support for FIt-SNE v1.2+
 - Fix for `Spatial*Plot` when running with interactive=TRUE
+- Set max for number of items returned by `Top` and remove duplicate items when balanced=TRUE
 
 ## [3.2.0] - 2020-07-15
 ### Added